requirements.txt
transformers>=4.37.2
torch>=2.2.0
streamlit
python-dotenv
